feat: add useWatchDependency hook to allow watched values in dep array
as the value of `watch` is cached, using them in dependency arrays do nothing (if they are arrays or objects), since their reference does not change every mutation.
To solve this, we need to use the useWatchDependency hook to subscribe and re-store the watched value in useState.

Not sure if this will cause perf issues, but for now it seems speedy enough.